+++ lib/bootstrap_form/form_builder.rb
@@ -156,12 +156,16 @@ module BootstrapForm
         div_class = ["custom-control", "custom-checkbox"]
         div_class.append("custom-control-inline") if layout_inline?(options[:inline])
         label_class = label_classes.prepend("custom-control-label").compact.join(" ")
+
+        label_options = { class: label_class }
+        label_options[:for] = options[:id] if options[:id].present?
+
         content_tag(:div, class: div_class.compact.join(" ")) do
           html = if options[:skip_label]
             checkbox_html
           else
-            # TODO: Notice we don't seem to pass the ID into the custom control.
-            checkbox_html.concat(label(label_name, label_description, class: label_class))
+            checkbox_html
+              .concat(label(label_name, label_description, label_options))
           end
           html.concat(generate_error(name)) if options[:error_message]
           html
@@ -170,14 +174,16 @@ module BootstrapForm
         wrapper_class = "form-check"
         wrapper_class += " form-check-inline" if layout_inline?(options[:inline])
         label_class = label_classes.prepend("form-check-label").compact.join(" ")
+
+        label_options = { class: label_class }
+        label_options[:for] = options[:id] if options[:id].present?
+
         content_tag(:div, class: wrapper_class) do
           html = if options[:skip_label]
             checkbox_html
           else
             checkbox_html
-              .concat(label(label_name,
-                            label_description,
-                            { class: label_class }.merge(options[:id].present? ? { for: options[:id] } : {})))
+              .concat(label(label_name, label_description, label_options))
           end
           html.concat(generate_error(name)) if options[:error_message]
           html
@@ -208,12 +214,16 @@ module BootstrapForm
         div_class = ["custom-control", "custom-radio"]
         div_class.append("custom-control-inline") if layout_inline?(options[:inline])
         label_class = label_classes.prepend("custom-control-label").compact.join(" ")
+
+        label_options = { value: value, class: label_class }
+        label_options[:for] = options[:id] if options[:id].present?
+
         content_tag(:div, class: div_class.compact.join(" ")) do
           html = if options[:skip_label]
             radio_html
           else
-            # TODO: Notice we don't seem to pass the ID into the custom control.
-            radio_html.concat(label(name, options[:label], value: value, class: label_class))
+            radio_html
+              .concat(label(name, options[:label], label_options))
           end
           html.concat(generate_error(name)) if options[:error_message]
           html
@@ -222,12 +232,16 @@ module BootstrapForm
         wrapper_class = "form-check"
         wrapper_class += " form-check-inline" if layout_inline?(options[:inline])
         label_class = label_classes.prepend("form-check-label").compact.join(" ")
+
+        label_options = { value: value, class: label_class }
+        label_options[:for] = options[:id] if options[:id].present?
+
         content_tag(:div, class: "#{wrapper_class}#{disabled_class}") do
           html = if options[:skip_label]
             radio_html
           else
             radio_html
-              .concat(label(name, options[:label], { value: value, class: label_class }.merge(options[:id].present? ? { for: options[:id] } : {})))
+              .concat(label(name, options[:label], label_options))
           end
           html.concat(generate_error(name)) if options[:error_message]
           html